fixed the verbose tkinter logs, moved stuff around
